# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Cirurgia Oncológica Campo Largo
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TimeSlot {
  /** Time in HH:MM format */
  time: string;
  available: boolean;
}

export interface AppointmentInput {
  patientName: string;
  patientEmail: string;
  patientPhone: string;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Time in HH:MM format */
  time: string;
  reason: string;
  /** @nullable */
  notes?: string | null;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export interface Appointment {
  id: number;
  patientName: string;
  patientEmail: string;
  patientPhone: string;
  date: string;
  time: string;
  reason: string;
  /** @nullable */
  notes?: string | null;
  status: AppointmentStatus;
  createdAt: string;
}

export type ChatHistoryItemRole =
  (typeof ChatHistoryItemRole)[keyof typeof ChatHistoryItemRole];

export const ChatHistoryItemRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatHistoryItem {
  role: ChatHistoryItemRole;
  content: string;
}

export interface ChatMessageInput {
  message: string;
  history?: ChatHistoryItem[];
}

export interface ChatResponse {
  reply: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetAvailableSlotsParams = {
  /**
   * Date in YYYY-MM-DD format
   */
  date: string;
};