# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Cirurgia Oncológica Campo Largo
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: appointments
    description: Appointment scheduling
  - name: chat
    description: Oncology AI chatbot
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /appointments/slots:
    get:
      operationId: getAvailableSlots
      tags: [appointments]
      summary: Get available appointment slots
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            description: Date in YYYY-MM-DD format
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"

  /appointments:
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Invalid input or slot unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /chat:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the oncology AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    TimeSlot:
      type: object
      required: [time, available]
      properties:
        time:
          type: string
          description: Time in HH:MM format
        available:
          type: boolean

    AppointmentInput:
      type: object
      required: [patientName, patientEmail, patientPhone, date, time, reason]
      properties:
        patientName:
          type: string
        patientEmail:
          type: string
        patientPhone:
          type: string
        date:
          type: string
          description: Date in YYYY-MM-DD format
        time:
          type: string
          description: Time in HH:MM format
        reason:
          type: string
        notes:
          type: ["string", "null"]

    Appointment:
      type: object
      required: [id, patientName, patientEmail, patientPhone, date, time, reason, status, createdAt]
      properties:
        id:
          type: integer
        patientName:
          type: string
        patientEmail:
          type: string
        patientPhone:
          type: string
        date:
          type: string
        time:
          type: string
        reason:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        createdAt:
          type: string

    ChatMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatHistoryItem"

    ChatHistoryItem:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string

    ChatResponse:
      type: object
      required: [reply]
      properties:
        reply:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string