# lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Cirurgia Oncológica Campo Largo
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentStatus } from "./appointmentStatus";

export interface Appointment {
  id: number;
  patientName: string;
  patientEmail: string;
  patientPhone: string;
  date: string;
  time: string;
  reason: string;
  /** @nullable */
  notes?: string | null;
  status: AppointmentStatus;
  createdAt: string;
}