# lib/api-zod/src/generated/types/appointmentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Cirurgia Oncológica Campo Largo
 * OpenAPI spec version: 0.1.0
 */

export interface AppointmentInput {
  patientName: string;
  patientEmail: string;
  patientPhone: string;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Time in HH:MM format */
  time: string;
  reason: string;
  /** @nullable */
  notes?: string | null;
}