# lib/api-zod/src/generated/types/timeSlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Cirurgia Oncológica Campo Largo
 * OpenAPI spec version: 0.1.0
 */

export interface TimeSlot {
  /** Time in HH:MM format */
  time: string;
  available: boolean;
}